+++ R/fit_survival_step.R
@@ -23,6 +23,7 @@
 #' # Testing dataset with just two treatment arms.
 #' library(scda)
 #' library(dplyr)
+#' library(rtables)
 #'
 #' adtte <- synthetic_cdisc_data("latest")$adtte
 #'
@@ -37,7 +38,7 @@
 #'     is_event = CNSR == 0
 #'   )
 #' labels <- c("ARM" = "Treatment Arm", "is_event" = "Event Flag")
-#' var_labels(adtte_f)[names(labels)] <- labels
+#' formatable::var_labels(adtte_f)[names(labels)] <- labels
 #'
 #' variables <- list(
 #'   arm = "ARM",