+++ R/h_response_biomarkers_subgroups.R
@@ -15,12 +15,12 @@
 #' library(rtables)
 #'
 #' adrs <- synthetic_cdisc_data("latest")$adrs
-#' adrs_labels <- var_labels(adrs)
+#' adrs_labels <- formatable::var_labels(adrs)
 #'
 #' adrs_f <- adrs %>%
 #'   filter(PARAMCD == "BESRSPI") %>%
 #'   mutate(rsp = AVALC == "CR")
-#' var_labels(adrs_f) <- c(adrs_labels, "Response")
+#' formatable::var_labels(adrs_f) <- c(adrs_labels, "Response")
 NULL
 
 #' @describeIn h_response_biomarkers_subgroups helps with converting the "response" function variable list
@@ -119,7 +119,7 @@ h_logistic_mult_cont_df <- function(variables,
       data.frame(
         # Dummy column needed downstream to create a nested header.
         biomarker = bm,
-        biomarker_label = var_labels(data[bm], fill = TRUE),
+        biomarker_label = formatable::var_labels(data[bm], fill = TRUE),
         n_tot = length(resp_vector),
         n_rsp = sum(resp_vector),
         prop = mean(resp_vector),
@@ -136,7 +136,7 @@ h_logistic_mult_cont_df <- function(variables,
   } else {
     data.frame(
       biomarker = variables$biomarkers,
-      biomarker_label = var_labels(data[variables$biomarkers], fill = TRUE),
+      biomarker_label = formatable::var_labels(data[variables$biomarkers], fill = TRUE),
       n_tot = 0L,
       n_rsp = 0L,
       prop = NA,