+++ R/h_response_subgroups.R
@@ -19,7 +19,7 @@
 #' library(rtables)
 #'
 #' adrs <- synthetic_cdisc_data("latest")$adrs
-#' adrs_labels <- var_labels(adrs)
+#' adrs_labels <- formatable::var_labels(adrs)
 #'
 #' adrs_f <- adrs %>%
 #'   filter(PARAMCD == "BESRSPI") %>%
@@ -30,7 +30,7 @@
 #'     ARM = fct_relevel(ARM, "B: Placebo"),
 #'     rsp = AVALC == "CR"
 #'   )
-#' var_labels(adrs_f) <- c(adrs_labels, "Response")
+#' formatable::var_labels(adrs_f) <- c(adrs_labels, "Response")
 NULL
 
 #' @describeIn h_response_subgroups helper to prepare a data frame of binary responses by arm.