+++ R/response_biomarkers_subgroups.R
@@ -18,12 +18,12 @@
 #' library(rtables)
 #'
 #' adrs <- synthetic_cdisc_data("latest")$adrs
-#' adrs_labels <- var_labels(adrs)
+#' adrs_labels <- formatable::var_labels(adrs)
 #'
 #' adrs_f <- adrs %>%
 #'   filter(PARAMCD == "BESRSPI") %>%
 #'   mutate(rsp = AVALC == "CR")
-#' var_labels(adrs_f) <- c(adrs_labels, "Response")
+#' formatable::var_labels(adrs_f) <- c(adrs_labels, "Response")
 NULL
 
 #' @describeIn response_biomarkers_subgroups prepares estimates for number of responses, patients and