+++ NEWS.md
@@ -12,6 +12,7 @@
 ### Miscellaneous
 
 * Move MMRM into a separate package `tern.mmrm`.
+* Removed `with_label`, `var_labels`, and `var_labels<-` to resolve conflict with `formatable`.
 
 
 # tern 0.7.6